+++ api/controllers/workinggroups.js
@@ -7,8 +7,8 @@ exports.createWG = async (req,res) => {
         if (isWG.length >= 1) {
             return res.status(409).json({message: "This Working Group already exist"});
         }
-        const actionDB = await WG.create(body)
-        res.json(actionDB)
+        const workgroupDB = await WG.create(body)
+        res.json(workgroupDB)
     } catch (error) {
         res.status(500).json({message: 'An error has occurred',error})
     }
@@ -16,8 +16,8 @@ exports.createWG = async (req,res) => {
 
 exports.getAllWG = async (req,res) => {
     try {
-        const actionDB = await WG.find()
-        res.json(actionDB)
+        const workgroupDB = await WG.find()
+        res.json(workgroupDB)
     } catch (error) {
         return res.status(400).json({mensaje: 'An error has occurred',error})
     }
@@ -27,12 +27,12 @@ exports.getWG = async (req,res) => {
     const _id = req.params.id
 
     try {
-        const actionDB = await WG.findById(_id,function(err,wg){
+        const workgroupDB = await WG.findById(_id,function(err,wg){
             if(wg==undefined){
                 return res.status(404).json({message:'Invalid ID'})
             }
         })
-        res.json(actionDB)
+        res.json(workgroupDB)
 
     } catch (error) {
         return res.status(400).json({message: 'An error has occurred',error })
@@ -44,8 +44,8 @@ exports.updateWG = async (req,res) => {
     const body = req.body
 
     try { 
-        const actionDB = await WG.findByIdAndUpdate(_id,body,{new:true})
-        res.json(actionDB)
+        const workgroupDB = await WG.findByIdAndUpdate(_id,body,{new:true})
+        res.json(workgroupDB)
     } catch (error) {
         return res.status(400).json({mensaje: 'An error has occurred',error})
     }
@@ -54,11 +54,11 @@ exports.updateWG = async (req,res) => {
 exports.deleteWG = async (req,res) => {
     const _id = req.params.id
     try {
-        const actionDB = await WG.findByIdAndDelete({_id})
-        if(!actionDB){
+        const workgroupDB = await WG.findByIdAndDelete({_id})
+        if(!workgroupDB){
             return res.status(400).json({message: 'ID not found',error})
         }
-       res.json(actionDB)
+       res.json(workgroupDB)
     } catch (error) {
         return res.status(400).json({mensaje: 'An error has occurred',error})
     }