+++ api/routes/comments.js
@@ -1,0 +1,6 @@
+// TODO: Comments Routes
+
+const express = require("express");
+const router = express.Router();
+
+module.exports = router;