+++ api/routes/configuration.js
@@ -1,0 +1,6 @@
+// TODO: Configuration Routes
+
+const express = require("express");
+const router = express.Router();
+
+module.exports = router;